Partial transformers v5 compatibility fixes
- Add exist_ok=True to all AutoModel/AutoModelForCausalLM.register() calls
  (v5 has built-in VibeVoice ASR which pre-registers some config classes)
- Add try/except for Qwen2TokenizerFast import (tokenization_*_fast.py
  files removed in v5, unified into single tokenization file)
- Accept **kwargs in all tie_weights() overrides (v5 passes recompute_mapping)
- Guard torch.linspace().item() in tokenizer init (v5 uses meta tensors
  during model construction)
- Guard sigmas.to("cpu") in DPM scheduler (same meta tensor issue)
- Add try/except for _prepare_generation_config (v5 removed use_model_defaults param)
- Add try/except for _prepare_cache_for_generation (handle varying arg counts)
- Add try/except for FeatureExtractionMixin import
- Add hasattr guard for batch_decode (unified into decode in v5)
- Add explicit torch_dtype in conversion script

Note: Full v5 compatibility is blocked by Qwen2 attention layer refactoring
in transformers v5 which restructured position embeddings and KV cache
handling. The longform inference code's generation loop is tightly coupled
to v4's Qwen2 internals. Tested working with transformers 4.51.3-4.57.3.

+++ vibevoice/scripts/convert_nnscaler_checkpoint_to_transformers.py
@@ -127,7 +127,7 @@ def convert_vibevoice_nnscaler_checkpoint_to_hf(
     
     # Verify the saved model can be loaded
     logger.info("Verifying saved model...")
-    loaded_model = VibeVoiceForConditionalGeneration.from_pretrained(pytorch_dump_folder_path)
+    loaded_model = VibeVoiceForConditionalGeneration.from_pretrained(pytorch_dump_folder_path, torch_dtype=torch.float32)
     logger.info("Model successfully loaded from saved checkpoint!")
 
 def main():
